Little reshuffling before the core gets more documentation
