+++ tests/test_beam.py
@@ -3,8 +3,9 @@ import numpy as np
 from numpy.testing import assert_equal
 from pySC.core.beam import bpm_reading, all_elements_reading, beam_transmission, generate_bunches
 from pySC.utils.sc_tools import SCgetOrds
-from pySC.core.classes import SimulatedComissioning
+from pySC.core.simulated_commissioning import SimulatedCommissioning
 from tests.test_at_wrapper import at_lattice
+
 
 def test_beam_transmission_basic(sc):
     max_turns, lost_fraction = beam_transmission(sc)
@@ -55,7 +56,7 @@ def test_generate_bunches_basic(sc):
 
 @pytest.fixture
 def sc(at_lattice):
-    SC = SimulatedComissioning(at_lattice)
+    SC = SimulatedCommissioning(at_lattice)
     SC.register_bpms(SCgetOrds(SC.RING, 'BPM'), CalError=5E-2 * np.ones(2),
                      Offset=500E-6 * np.ones(2),
                      Noise=10E-6 * np.ones(2),
