+++ tests/test_simulated_commissioning.py
@@ -4,12 +4,12 @@ import pytest
 import numpy as np
 from numpy.testing import assert_allclose, assert_equal
 import at
-from pySC.core.classes import SimulatedComissioning
+from pySC.core.simulated_commissioning import SimulatedCommissioning
 from pySC.utils.sc_tools import SCgetOrds
 
 
 def test_simulated_commissioning_data_structure(at_cell):
-    sc = SimulatedComissioning(at_cell)
+    sc = SimulatedCommissioning(at_cell)
     sc.RING[9].PolynomB[1] = 1
     assert sc.RING[10].PolynomB[1] == 1.2
     assert sc.IDEALRING[9].PolynomB[1] == 1.2
@@ -31,7 +31,7 @@ def test_simulated_commissioning_data_structure(at_cell):
 
 
 def test_register_bpms(at_cell):
-    sc = SimulatedComissioning(at_cell)
+    sc = SimulatedCommissioning(at_cell)
     bpm_dict = dict(CalError=5E-2 * np.ones(2), Offset=500E-6 * np.ones(2),
                     Noise=10E-6 * np.ones(2), NoiseCO=1E-6 * np.ones(2), Roll=1E-3)
     sc.register_bpms(np.array([8, 21]), CalError=5E-2 * np.ones(2), Offset=500E-6 * np.ones(2),
@@ -45,7 +45,7 @@ def test_register_bpms(at_cell):
 
 
 def test_register_cavities(at_cell):
-    sc = SimulatedComissioning(at_cell)
+    sc = SimulatedCommissioning(at_cell)
     rf_dict = dict(FrequencyOffset=5E3, VoltageOffset=5E3, TimeLagOffset=0.5)
     sc.register_cavities(np.array([0, 0, 0]), FrequencyOffset=5E3, VoltageOffset=5E3, TimeLagOffset=0.5)
     assert_equal(sc.ORD.RF, np.array([0]))
@@ -53,7 +53,7 @@ def test_register_cavities(at_cell):
 
 
 def test_register_magnets(at_cell):
-    sc = SimulatedComissioning(at_cell)
+    sc = SimulatedCommissioning(at_cell)
     mag_dict = dict(CalErrorB=np.array([5E-2, 1E-3]), MagnetOffset=200E-6 * np.array([1, 1, 0]),)
     indices = SCgetOrds(sc.RING, "Q")
     sc.register_magnets(np.repeat(indices, 2), CalErrorB=np.array([5E-2, 1E-3]), MagnetOffset=200E-6 * np.array([1, 1, 0]),)
@@ -66,7 +66,7 @@ def test_register_magnets(at_cell):
 
 
 def test_register_supports(at_cell):
-    sc = SimulatedComissioning(at_cell)
+    sc = SimulatedCommissioning(at_cell)
     mag_dict = dict(SectionOffset=200E-6 * np.array([1, 1, 0]),)
     indices = np.vstack((SCgetOrds(sc.RING, "SF"), SCgetOrds(sc.RING, "SD")))
     sc.register_supports(np.repeat(indices, 2, axis=1), "Section",  Offset=200E-6 * np.array([1, 1, 0]),)
@@ -79,8 +79,9 @@ def test_register_supports(at_cell):
     assert_equal(sc.SIG.Support[indices[0, 1]], mag_dict)
     assert_equal(sc.SIG.Support[indices[1, 1]], update_dict2)
 
+
 def test_register_bad_keywords(at_cell):
-    sc = SimulatedComissioning(at_cell)
+    sc = SimulatedCommissioning(at_cell)
     with pytest.raises(ValueError) as e_info:
         sc.register_bpms(np.array([8, 21]), BadKeyword=5E-2 * np.ones(2),)
     assert "BadKeyword" in str(e_info.value)
@@ -100,7 +101,7 @@ def test_register_bad_keywords(at_cell):
 
 def test_apply_errors(at_cell):
     np.random.seed(123)
-    sc = SimulatedComissioning(at_cell)
+    sc = SimulatedCommissioning(at_cell)
     sc.register_bpms(np.array([8, 21]), CalError=5E-2 * np.ones(2), Offset=500E-6 * np.ones(2),
                      Noise=10E-6 * np.ones(2), NoiseCO=1E-6 * np.ones(2), Roll=1E-3)
     sc.register_cavities(np.array([0]), FrequencyOffset=5E3, VoltageOffset=5E3, TimeLagOffset=0.5)
@@ -149,8 +150,6 @@ def test_apply_errors(at_cell):
 
         assert_allclose(getattr(sc.RING[10], attr), value, rtol=2e-5)
         assert not hasattr(sc.IDEALRING[10], attr)
-
-
 
 
 @pytest.fixture
